+++ docs/changelog.md
@@ -41,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Install a component while running `components install [name]` with no `--idtf` flag
+- Exception while running `components search --explanation` with an empty substring
 - Use nrel_authors instead of nrel_author
 
 ### Removed